requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
comtypes==1.1.11
idna==3.3
lxml==4.8.0
numpy==1.22.2
pandas==1.4.1
PyGLM==2.5.7
PyOpenGL==3.1.6
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
PySide2==5.15.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pywinauto==0.6.8
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
shiboken2==5.15.2.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
